allow custom processing of markdown content
Stores markdown content between code blocks, and executes `mdsh-md-compile`
function (if defined) passing the markdown content and the language tag
of the following code block into it. The function is executed immediately
before the language hooks for the following code block.

Also, pass the markdown content as last parameter to the `mdsh-compile-lang`
function in case someone whats to us the content when generating
code (for example to add logging statements based on the content).